#206 冗長な実装だったため修正

+++ app/controllers/comments_controller.rb
@@ -1,6 +1,5 @@
 class CommentsController < ApplicationController
   before_action :authenticate_user!
-  before_action :limit_user, only: [:destroy]
 
   def create
     Comment.create(comment_params)
@@ -9,7 +8,9 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = Comment.find(params[:id])
-    comment.destroy
+    if current_user.same_user?(comment.user_id)
+      comment.destroy
+    end
     redirect_back(fallback_location: root_path)
   end
 
@@ -17,12 +18,5 @@ class CommentsController < ApplicationController
 
   def comment_params
     params.require(:comment).permit(:content).merge(post_id: params[:post_id], user_id: current_user.id)
-  end
-
-  def limit_user
-    comment = Comment.find(params[:id])
-    unless current_user.same_user?(comment.user_id)
-      redirect_back(fallback_location: root_path)
-    end
   end
 end
